Issue 73 step 2: extract poll_binding into worker/lifecycle.rs
The 289-LOC poll_binding orchestrator was the largest function in
afxdp/mod.rs and the only function with structural reason to live
elsewhere — it is the per-poll BindingWorker lifecycle, called once
per binding per worker-loop iteration from worker/mod.rs:worker_loop.

Mechanical move (no production logic touched):
- Extracted poll_binding (and its `MAX_RX_BATCHES_PER_POLL >= 1`
  compile-time guard prelude) to a new userspace-dp/src/afxdp/worker/lifecycle.rs
- Visibility bumped from private to `pub(super)` so worker/mod.rs can
  import it
- worker/mod.rs adds `mod lifecycle;` + `use lifecycle::poll_binding;`
- afxdp/mod.rs drops 298 lines (775 → 475 LOC)

Lifecycle.rs uses `use super::*;` (worker scope, which itself does
`use super::*;` to pull from afxdp/mod.rs scope). All types, helpers,
and constants poll_binding references resolve through that chain.

cargo build --release: 92 warnings baseline, no errors.
cargo test --release: 874 passed, 0 failed.

afxdp/mod.rs is now 475 LOC and contains only:
- imports and module declarations (1-302)
- BatchCounters struct + impl (303-394)
- mod poll_descriptor / session_delta / neighbor_dispatch / disposition
  / forward_request declarations + use clauses (390-440)
- PendingForwardHints (442-453)
- binding_by_index_mut + find_target_binding_mut helpers (442-475)
